New class for contiguous dynamic multi-dimensional arrays
The key word here is "contiguous". To fill ROOT TBranches is much much simpler
to use static C-arrays because they are allocated in a contiguous chunck of
memory. But this comes with all the downsides of using old C-arrays. On the
other hand, the dictionary generation for nested classes in ROOT doesn't seem to
support triple nested classes, so we cannot something like
<vector<vector<vector<T>>> to fill a TBranch. Waiting for proper nested classes
support in the ROOT dictionary generator, I created a class that behaves like a
2D or 3D dinamical array with the constraint that the memory is contiguously
allocated. The class automatically takes care of the memory management and
checks for out of range errors. If needed the 1D-subarrays can be wrapped in a
std::vector through the wrapArrayInVector function.