New wgNotInitialized WAGASCI exception
This exception is thrown when an object has been constructed but not yet
initialized. I known that this should never happen for the RAII principle but
this is unavoidable if we want to declare and initialized objects in different
places and been able to use the "." instead of the "->". For the time being this
only applies to the ContiguousVectors classes.