Use double instead of float to store the decoded data
Apparently if you don't use double for the histogram variables the software
crashes. It is not a critical issue so just switched to double variables
everywhere even if most probably we will never need double precision. Probably I
will take a deeper look into this issue in the future. Probably not.