Solved a nasty bug of the Logger
In the Logger source file (wgTool.cc) a global object (Log) of the class Logger
is defined. That object is used for all the logging. But if multiple compilation
units are compiled separately and then linked together multiple copies of that
global object are created. So far so good. The problem is that when the
constructor and destructor of those objects are called, the same files are
opened and closed multiple times, leading to invalid free() errors and
segmentation faults (visible with valgrind for example). The solution I chose is
to compile all the objects that use the Logger class as a single compilation
units (no libraries) so that the global object is defined only ones. Before I
was compiling the process programs as "main" program  + library, so that the
object was defined twice. A more long term solution would be to implement a
totally different logging system (maybe in a completely separate process).