Debugged wgDecoder
A line of the raw data is made up of two bytes. The software can recover corrupted data if one or more lines are missing. But if the raw data was missing a single byte, the offset would make all the successive reads fail. Then the software would try to read the whole file over and over again without ever finding a header or a trailer. The stack of the program would fill up and there you have a segmentation fault. Increasing the stack would only postpone the inevitable. The solution is locate and discard the corrupted byte.